requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
codecovopentelem @ git+ssh://git@github.com/codecov/opentelem-python.git@df1e241927d4794fc0e24f96431f0fb730edac21
coverage==6.2
decorator==5.1.0
Deprecated==1.2.13
distlib==0.3.4
filelock==3.4.2
Flask==2.0.2
idna==3.3
ipdb==0.13.9
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
opentelemetry-api==1.8.0
opentelemetry-instrumentation==0.27b0
opentelemetry-instrumentation-flask==0.27b0
opentelemetry-instrumentation-requests==0.27b0
opentelemetry-instrumentation-wsgi==0.27b0
opentelemetry-sdk==1.8.0
opentelemetry-semantic-conventions==0.27b0
opentelemetry-util-http==0.27b0
parso==0.8.3
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.10.0
requests==2.26.0
six==1.16.0
stevedore==3.5.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.7
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
